requirements.txt
openai
chromadb
psycopg2-binary
mss
Pillow
pydirectinput
python-dotenv
